Add document for `glzip`
