+++ docs/man/man1/glzip.1.adoc
@@ -43,7 +43,7 @@ _FILE_::
 
   Keep input files.
 
-*-dictionary-size*::
+*-dictionary-size* _BYTE_::
 
   Set dictionary size in bytes.
